Initial implementation
Extracted from a PR against [ember-cli-rails].

`EmberCLI::Deploy` takes a `namespace` (the name of your app declared
in your initializer) and handles all interaction with the Redis
instance.

This is great for `staging` and `production` deploys, but introduces an
extra step in the feedback loop during development.

Luckily, `EmberCLI::Deploy` also accepts an `index_html` override, which
will replace the call to the Redis instance. This allows integration
with the normal `ember-cli-rails` workflow:

[ember-cli-rails]: https://github.com/rwz/ember-cli-rails/pull/172